Use BigInteger for stake amounts
With this change, WAL-12 is fixed – the minimum stake amount is not
checked if not changing the amount while updating the status.

+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/baker/BakerRegistrationConfirmationActivity.kt
@@ -15,7 +15,6 @@ import com.concordium.wallet.databinding.ActivityBakerRegistrationConfirmationBi
 import com.concordium.wallet.ui.account.accountdetails.AccountDetailsActivity
 import com.concordium.wallet.ui.bakerdelegation.common.BaseDelegationBakerActivity
 import com.concordium.wallet.util.UnitConvertUtil
-import com.concordium.wallet.util.toBigInteger
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import java.math.BigInteger
 
@@ -81,7 +80,7 @@ class BakerRegistrationConfirmationActivity : BaseDelegationBakerActivity(
 
             UPDATE_BAKER_KEYS -> {
                 viewModel.bakerDelegationData.amount =
-                    viewModel.bakerDelegationData.account?.accountBaker?.stakedAmount.toBigInteger()
+                    viewModel.bakerDelegationData.account?.accountBaker?.stakedAmount
                 updateViewsUpdateBakerKeys()
             }
